made it compileable for all platforms.
[git-p4: depot-paths = "//depot/qt/main/": change = 66996]

+++ tools/designer/designer/metadatabase.cpp
@@ -1117,8 +1117,8 @@ QString MetaDataBase::extractVariableName( const QString &name )
     QString n = name.right( name.length() - name.findRev( ' ' ) - 1 );
     if ( n[ 0 ] == '*' || n[ 0 ] == '&' )
 	n[ 0 ] = ' ';
-    if ( n[ n.length() - 1 ] == ';' )
-	n[ n.length() - 1 ] = ' ';
+    if ( n[ (int)n.length() - 1 ] == ';' )
+	n[ (int)n.length() - 1 ] = ' ';
     return n.simplifyWhiteSpace();
 }
